+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Clarify FramesTracker log message (#3570)
 - Fix rare battery breadcrumbs crash (#3582)
 - Fix synchronization issue in FramesTracker (#3571)
+- Fix FileManager logs info instead of error when a path doesn't exist (#3594)
 
 ## 8.19.0